requirements.txt
Django==5.1.1
nltk==3.9.1
sentence-transformers==3.1.1
chromadb==0.5.5
google-generativeai==0.8.3
python-decouple==3.8